app/backend/requirements.txt
faiss-cpu 
torch 
transformers 
sentence_transformers
sentencepiece
requests
pandas 
numpy
PyMuPDF 
Pillow 
opencv-python 
easyocr[ru] 
streamlit